requirements.txt
# FastAPI and Server
fastapi
uvicorn[standard]
python-multipart

# PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0

# Training (optional - only needed for retraining)
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Scraping (optional - only needed for dataset collection)
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
